#41 refactoring shop code
