+++ GoodGames/game/models.py
@@ -2,3 +2,10 @@ from django.db import models
 from django.contrib.auth.models import User
 
 # Create your models here.
+
+class buyNow(models.Model):
+    name = models.CharField("Full name", max_length=1024)
+    address1 = models.CharField("Address line 1", max_length=1024)
+    address2 = models.CharField("Address line 2", max_length=1024, blank=True, null=True)
+    zip_code = models.CharField("ZIP", max_length=12)
+    city = models.CharField("City", max_length=1024)
