+++ GoodGames/game/urls.py
@@ -4,4 +4,5 @@ from . import views
 urlpatterns = [
     path('', views.home, name='home'),
     path('shop/', views.shop, name='shop'),
+    path('buy/', views.buy_view, name='buy'),
 ]
