+++ GoodGames/users/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth.models import User
 from django.contrib.auth.forms import UserCreationForm
-from .models import buyNow
 from django.forms import ModelForm
 class UserRegisterForm(UserCreationForm):
     email = forms.EmailField()
@@ -15,10 +14,3 @@ class UserRegisterForm(UserCreationForm):
         fields = ['username', 'email', 'password1', 'password2']
 
 
-
-class buyNowForm(ModelForm):
-    class Meta:
-        model = buyNow
-        fields = [
-            'name', 'address1', 'address2', 'zip_code', 'city'
-            ]
