+++ GoodGames/users/models.py
@@ -1,8 +1,20 @@
 from django.db import models
+from django.contrib.auth.models import User
 # Create your models here.
-class buyNow(models.Model):
-    name = models.CharField("Full name", max_length=1024)
-    address1 = models.CharField("Address line 1", max_length=1024)
-    address2 = models.CharField("Address line 2", max_length=1024, blank=True, null=True)
-    zip_code = models.CharField("ZIP", max_length=12)
-    city = models.CharField("City", max_length=1024)
+
+
+class Profile(models.Model):
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+
+    bio = models.TextField(default=None, null=True)
+
+    balance = models.IntegerField(default=100000)
+
+    address = models.CharField(max_length=100, default=None, null=True)
+    city = models.CharField(max_length=100, default=None, null=True)
+    state = models.CharField(max_length=2, default=None, null=True)
+    zip_code = models.CharField(max_length=5, default=None, null=True)
+    
+    
+
+
